refactor: move variables out of types
