+++ great_expectations/data_context/store/inline_store_backend.py
@@ -1,11 +1,11 @@
 import logging
 from typing import Any, List, Optional, Tuple
 
-from great_expectations.data_context.store.store_backend import StoreBackend
-from great_expectations.data_context.types.base import DataContextConfig
-from great_expectations.data_context.types.data_context_variables import (
+from great_expectations.data_context.data_context_variables import (
     DataContextVariableSchema,
 )
+from great_expectations.data_context.store.store_backend import StoreBackend
+from great_expectations.data_context.types.base import DataContextConfig
 from great_expectations.exceptions.exceptions import StoreBackendError
 from great_expectations.util import filter_properties_dict
